TLV ingestion improvements

+++ include/ZMQParserInterface.h
@@ -44,8 +44,8 @@ class ZMQParserInterface : public ParserInterface {
   void preprocessFlow(ParsedFlow *flow, NetworkInterface *iface);
   bool getKeyId(char *sym, u_int32_t * const pen, u_int32_t * const field) const;
   void addMapping(const char *sym, u_int32_t num, u_int32_t pen = 0);
-  bool parsePENZeroField(ParsedFlow * const flow, u_int32_t field, const char * const value, u_int32_t ivalue) const;
-  bool parsePENNtopField(ParsedFlow * const flow, u_int32_t field, const char * const value, u_int32_t ivalue) const;
+  bool parsePENZeroField(ParsedFlow * const flow, u_int32_t field, const char * const value, u_int64_t ivalue) const;
+  bool parsePENNtopField(ParsedFlow * const flow, u_int32_t field, const char * const value, u_int64_t ivalue) const;
   static bool parseContainerInfo(json_object *jo, ContainerInfo * const container_info);
   bool parseNProbeMiniField(ParsedFlow * const flow, const char * const key, const char * const value, json_object * const jvalue) const;
   void parseSingleJSONFlow(json_object *o, u_int8_t source_id, NetworkInterface *iface);
